requirements.txt
# requirements.txt

# FastAPI ve ilgili paketler
fastapi==0.104.1
uvicorn==0.24.0
pydantic[email]==2.4.2  # email validator dahil
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.1.0.post1

# Veritabanı
pymysql==1.1.0
SQLAlchemy==2.0.23
alembic==1.12.1
cryptography==41.0.5

# Test araçları
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1

# Geliştirici araçları
black==23.10.1
flake8==6.1.0
isort==5.12.0

# JWT ve güvenlik
python-jose[cryptography]==3.3.0